+++ client/src/component/clusterinfo/api.js
@@ -1,9 +1,23 @@
 'use strict';
 //@flow
 
+import type { AgentDetails } from './action';
 import type { ClusterMeta } from '../dashboard/action';
 import type { ClusterDiscovery } from './action';
 import type { Event } from '../events/action';
+
+
+export function fetchAgents(cluster_id: string): Promise<Array<AgentDetails>> {
+  let url = `/api/cluster/${cluster_id}/agents`;
+  return fetch(url).then((response) => {
+    return response.json().then((body) => {
+      if (!response.ok) {
+        throw Error('Fetch error: ' + body.error);
+      }
+      return body;
+    });
+  });
+}
 
 
 export function fetchDiscovery(cluster_id: string): Promise<ClusterDiscovery> {