+++ client/src/component/clusterinfo/index.js
@@ -15,6 +15,7 @@ import type { ClusterMeta } from '../dashboard/action';
 import type { ClusterInfoStore } from './store';
 import { fetchMeta } from './action';
 import { NoDataIcon } from '../icons';
+import Agents from './agents';
 import Discovery from './discovery';
 import Events from './events';
 import Overview from './overview';
@@ -60,6 +61,7 @@ export class InnerCluster extends React.Component<Props> {
     if (root.endsWith('/')) {
       root = root.substring(0, root.length - 1);
     }
+    const agents = `${root}/agents`;
     const discovery = `${root}/discovery`;
     const events = `${root}/events`;
     const overview = `${root}/overview`;
@@ -68,6 +70,7 @@ export class InnerCluster extends React.Component<Props> {
         <Route exact path={root} component={OverviewRedirect} />
         <Route path={overview} render={props => <Overview cluster={meta} {...props} />} />
         <Route path={discovery} component={Discovery} />
+        <Route path={agents} component={Agents} />
         <Route path={events} component={Events} />
       </div>
     );
@@ -78,6 +81,7 @@ export class InnerCluster extends React.Component<Props> {
     if (root.endsWith('/')) {
       root = root.substring(0, root.length - 1);
     }
+    const agents = `${root}/agents`;
     const discovery = `${root}/discovery`;
     const events = `${root}/events`;
     const overview = `${root}/overview`;
@@ -85,6 +89,7 @@ export class InnerCluster extends React.Component<Props> {
       <div className="cluster-tabs">
         <NavLink className="tab" to={overview}>Overview</NavLink>
         <NavLink className="tab" to={discovery}>Discovery</NavLink>
+        <NavLink className="tab" to={agents}>Agents</NavLink>
         <NavLink className="tab" to={events}>Events</NavLink>
       </div>
     );