+++ client/src/component/dashboard/__tests__/events.js
@@ -38,11 +38,11 @@ describe('Dashboard', () => {
 
     test('renders with events', () => {
       let events = [{
-        data: "SOME DATA",
+        data: {"SOME": "DATA"},
         event: "AGENT_DOWN",
         timestamp: "2018-04-28T17:57:24.540480643Z"
       }, {
-        data: "SOME DATA",
+        data: {"SOME": "DATA"},
         event: "AGENT_RECOVER",
         timestamp: "2018-04-28T17:50:15.170187929Z"
       }];